feat: [OptimizationChart] moved chart from TestRunsChartPage into TestRunsChart
